# Conflicts:
#	app/src/main/java/com/example/capstone_project/EventForms.java
#	app/src/main/java/com/example/capstone_project/MainMenu.java
#	app/src/main/java/com/example/capstone_project/utils/EventServiceManager.java

+++ app/src/main/java/com/example/capstone_project/EventDetails.java
@@ -4,7 +4,9 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
+import android.widget.Toast;
 
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
@@ -22,9 +24,13 @@ public class EventDetails extends AppCompatActivity {
     private TextView eventTitle;
     private TextView eventDescription;
     private TextView eventStartDate;
+
+    private TextView registerAttendeeButton;
+    private TextView unregisterAttendeeButton;
     private TextView verifyAttendeeButton;
     private TextView editDetailsButton;
     private TextView deleteEventButton;
+
     private TextView numAttendeesRegistered;
     private TextView numRemainingSlots;
     private TextView numTotalRevenue;
@@ -32,6 +38,7 @@ public class EventDetails extends AppCompatActivity {
     private RecyclerView attendeeList;
 
     private Event event;
+    private String[] attendeeListArray;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -41,9 +48,13 @@ public class EventDetails extends AppCompatActivity {
         eventTitle = findViewById(R.id.eventDetailName);
         eventDescription = findViewById(R.id.eventDetailDescription);
         eventStartDate = findViewById(R.id.eventDetailStartDate);
+
+        registerAttendeeButton = findViewById(R.id.eventDetails_registerAttendant_button);
+        unregisterAttendeeButton = findViewById(R.id.eventDetails_unRegisterAttendant_button);
         verifyAttendeeButton = findViewById(R.id.eventDetails_verifyAttendant_button);
         editDetailsButton = findViewById(R.id.eventDetails_editDetails_button);
         deleteEventButton = findViewById(R.id.eventDetails_deleteEvent_button);
+
         numAttendeesRegistered = findViewById(R.id.eventDetails_attendeesRegistered);
         numRemainingSlots = findViewById(R.id.eventDetails_remainingSlots);
         numTotalRevenue = findViewById(R.id.eventDetails_totalRevenue);
@@ -51,6 +62,13 @@ public class EventDetails extends AppCompatActivity {
         attendeeList = findViewById(R.id.eventDetails_attendeeListView);
 
         // TODO: register and unregister attendee
+
+        unregisterAttendeeButton.setOnClickListener(v -> {
+            Intent intent = new Intent(v.getContext(), UnregisterAttendee.class);
+            intent.putExtra("SELECTED_EVENT_ID", event.getEventId());
+            intent.putExtra("ATTENDEES", attendeeListArray);
+            v.getContext().startActivity(intent);
+        });
 
         verifyAttendeeButton.setOnClickListener(v -> {
             Intent intent = new Intent(v.getContext(), VerifyAttendee.class);
@@ -61,13 +79,25 @@ public class EventDetails extends AppCompatActivity {
         editDetailsButton.setOnClickListener(v -> {
             Intent intent = new Intent(v.getContext(), EventForms.class);
             intent.putExtra("SELECTED_EVENT_ID", event.getEventId());
+            intent.putExtra("ATTENDEES", attendeeListArray);
             v.getContext().startActivity(intent);
         });
 
         deleteEventButton.setOnClickListener(n -> {
-            // TODO: should probably add confirmation to delete
-            EventServiceManager.getInstance().deleteEvent(event.getEventId());
-            finish();
+            // DONE: should probably add confirmation to delete
+            AlertDialog.Builder popup = new AlertDialog.Builder(this);
+            popup.setTitle("Delete Event");
+            popup.setMessage("Are you sure you want to delete " + event.getName() + "?");
+            popup.setPositiveButton("Yes", (dialog, which) -> {
+                Toast.makeText(this, "Deleted " + event.getName(), Toast.LENGTH_SHORT).show();
+                EventServiceManager.getInstance().deleteEvent(event.getEventId());
+                finish();
+            });
+            popup.setNegativeButton("No", (dialog, which) -> {
+                dialog.dismiss();
+            });
+            AlertDialog dialog = popup.create();
+            dialog.show();
         });
 
     }
@@ -76,7 +106,7 @@ public class EventDetails extends AppCompatActivity {
     protected void onResume() {
         super.onResume();
         event = EventServiceManager.getInstance().getEventFromId(getIntent().getStringExtra("SELECTED_EVENT_ID"));
-        String[] attendeeListArray = EventServiceManager.getInstance().getAttendeeNames(event.getEventId());
+        attendeeListArray = EventServiceManager.getInstance().getAttendeeNames(event.getEventId());
 
         eventTitle.setText(event.getName());
 
@@ -102,7 +132,7 @@ public class EventDetails extends AppCompatActivity {
             noAttendeeText.setVisibility(View.GONE);
         }
 
-        AttendeeListAdapter attendeeListAdapter = new AttendeeListAdapter(attendeeListArray);
+        AttendeeListAdapter attendeeListAdapter = new AttendeeListAdapter(event.getEventId(), attendeeListArray, 0);
         attendeeList.setLayoutManager(new LinearLayoutManager(this));
         attendeeList.setAdapter(attendeeListAdapter);
     }
